found a bug in my code. I wrote it so any user can delete any organization.Even if another user is already associated with the organization. This breaks the code if another user is already asscoiated with the org. So I coded it out an if statement that prevents that.

+++ adnat (backend)/app/controllers/organisations_controller.rb
@@ -16,8 +16,11 @@ class OrganisationsController < ApplicationController
 
     def destroy
         old_org = Organisation.find(params[:id])
-        old_org.destroy
-        head :no_content 
+        if User.all.pluck(:organisation_id).include?(old_org.id)
+            puts "Other Users are associated with this organization", status: :ok
+        else
+            old_org.destroy
+        end
     end
 
     private
